4_Implementación/backend/requirements.txt
flask
python-dotenv
flask-restful
flask-sqlalchemy
flask-jwt-extended
flask_restx
flask_login
werkzeug
flask-mail
mysqlclient
marshmallow-sqlalchemy
flask_cors
flask_marshmallow